% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.to.r
\name{d.to.r}
\alias{d.to.r}
\title{Convert cohen's d into r}
\usage{
d.to.r(d, p = 0.5)
}
\arguments{
\item{d}{cohen's d}

\item{p}{the proportion of individuals in group 1}
}
\description{
Convert cohen's d into r
}
\examples{
d.to.r(1.1, p=.3)
}
\author{
Dustin Fife
}
\seealso{
\code{\link{r.to.d}}
}

