% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc4.R
\name{dc4}
\alias{dc4}
\title{Estimate Cohen's d Using Case IV}
\usage{
dc4(data, grouping.var, rxxa = NULL)

dc4(data, grouping.var, rxxa = NULL)
}
\arguments{
\item{data}{The dataset containing missing values}

\item{grouping.var}{The categorical variable that indicates group membership}

\item{rxxa}{The reliability at the applicant (population) level}

\item{predictor.var}{The variable the researcher wishes to estimate subgroup differences for}

\item{rxxi}{The reliability at the incumbent (selected) level}

\item{sd.x.pop}{The population standard deviation of x}

\item{grouping.var}{The name of the grouping variable (surrounded by quotes)}

\item{rxxa}{Optional. The value of the reliability at the population level (to correct for unreliability)}
}
\description{
Estimate Cohen's d Using Case IV

Estimate Cohen's d Using the EM Algorithm
}
\examples{
grouping.var = rep(c(1:2), times=50)
predictor.var = 30 + 5*grouping.var + rnorm(100, 0, 5)
dc4(predictor.var, grouping.var, rxxa=.8, rxxi=.7, sd.x.pop=7)
}
\author{
Dustin Fife

Dustin Fife
}
