% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc4.R
\name{dpl}
\alias{dpl}
\title{Correct Cohen's d Using the Pearson Lawley}
\usage{
dpl(selected.cov, pop.cov, Q, grouping.var, outcome.var)
}
\arguments{
\item{selected.cov}{The selected variance/covariance matrix of all variables}

\item{pop.cov}{The population variance/covariance matrix of all variables used for selection}

\item{Q}{The proportion of applicants belonging to the minority group}

\item{grouping.var}{The index of the grouping variable (i.e., which column indicates the parameters for the grouping variable?)}

\item{outcome.var}{The index of the outcome variable (i.e., which column indicates the parameters for the outcome variable?)}
}
\value{
The pearson lawley estimate of d
}
\description{
Correct Cohen's d Using the Pearson Lawley
}
\author{
Dustin Fife
}
