% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeDV.R
\name{maximizeDV}
\alias{maximizeDV}
\title{Estimate the maximum value of a criterion under optimal selection}
\usage{
maximizeDV(formula, data, method = c("regression", "logistic",
  "poisson"), imputations = 30, plot = TRUE)
}
\arguments{
\item{formula}{A equation of the form y~x1 + x2... The Y is the value to be maximized by optimizing selection}

\item{data}{The dataset}

\item{method}{One of the following: "regression", "logistic", or "poisson", which specifies the type of model to fit
to the criterion}

\item{imputations}{Number of imputations. Defaults to 30.}

\item{plot}{Should a plot be returned? Defaults to TRUE.}
}
\description{
Estimate the maximum value of a criterion under optimal selection
}
