% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mv.correction}
\alias{mv.correction}
\title{Estimate the unrestricted covariance matrix using the Pearson-Lawley's Multivariate Correction}
\usage{
mv.correction(data, p, v.pp)
}
\arguments{
\item{data}{Either a data matrix containing the restricted dataset, or a variance covariance matrix of the selected covariances.
The variables used for selected (the Xs) must be in the first \code{p}
columns, while the "incidental" variables are in the last n-p:p columns}

\item{p}{the number of columns used for selection}

\item{v.pp}{The variance-covariance matrix of the population selection variables}
}
\value{
\item{}{}
}
\description{
Estimate the unrestricted covariance matrix using the Pearson-Lawley's Multivariate Correction
}
\examples{
# do a simulation to demonstrate unbiasedness of it
	#### specify population correlation matrix
	cor.mat = matrix(c(
		1, .2, .4, .5,
		.2, 1, .3, .6,
		.4, .3, 1, .4,
		.5, .6, .4, 1), ncol=4)

		#### sample from that matrix
	d = data.frame(mvrnorm(100000, mu=rep(0, times=ncol(cor.mat)), Sigma=cor.mat)	)
	names(d) = c("X1", "X2", "Y1", "Y2")

		#### restrict the data according to X1 and X2
	d.r = d[which((d$X1)>0 & d$X2>0),]

	cor.mat.corrected = mv.correction(data=d.r, p=2, v.pp = cor.mat[1:2, 1:2])

# compare the two estimates
	cor.mat
	round(cor.mat.corrected, digits=4)
}
\author{
Dustin Fife
}
\seealso{
\code{\link{MASS}}
}

