% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximizeDV.R
\name{print.selection}
\alias{print.selection}
\title{Print selection Summary}
\usage{
\method{print}{selection}(x, ...)
}
\arguments{
\item{x}{a selection object}

\item{...}{ignored}
}
\description{
Print selection Summary
}
