% Generated by roxygen2 (4.0.2): do not edit by hand
\name{r.to.d}
\alias{r.to.d}
\title{Convert r into cohen's d}
\usage{
r.to.d(r, p = 0.5)
}
\arguments{
\item{r}{the correlation coefficient}

\item{p}{the proportion of individuals in group 1}
}
\description{
Convert r into cohen's d
}
\examples{
r.to.d(.3, p=.3)
}
\author{
Dustin Fife
}
\seealso{
\code{\link{d.to.r}}
}

