% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.correction.R
\name{rel.correction}
\alias{rel.correction}
\title{Corrects a correlation for unreliability}
\usage{
rel.correction(rxyi, rxxi = NULL, ryyi = NULL)
}
\arguments{
\item{rxyi}{the obtained restricted correlation coefficient (obtained from
the incumbents).}

\item{rxxi}{the obtained reliability of x in the sample}

\item{ryyi}{the obtained reliability of y in the sample}
}
\value{
returns a scalar that is the corrected correlation.
}
\description{
\code{rel.correction} corrects obtained correlation for unreliability.
}
\examples{

	rel.correction(rxyi=.3, rxxi=.8)
	rel.correction(rxyi=.3, rxxi=.8, ryyi=.6)
	rel.correction(rxyi=.3, ryyi=.6)		
}
\author{
Dustin Fife
}

